+++ RELEASE_NOTES.md
@@ -10,9 +10,11 @@
     * `loadTable`
     * `loadTheme`
   * Added helper methods to `index.js` for writing M8 files/types
+    * `dumpInstrument`
     * `dumpScale`
     * `dumpTheme`
   * Added support for reading table data from an `Instrument` file
+  * Added `M8FileWriter` to `lib/types/`
   * Moved `M8FileReader` from `index.js` to `lib/types/`
   * Moved all classes from `lib/types.js` to `lib/types/` into their own files
     * `lib/types/Chain.js`
@@ -35,6 +37,7 @@
     * `None`
     * `Sampler`
     * `Wavsynth`
+  * Updated `lib/types/M8FileReader.js` to no longer keep track of skipped bytes and instead return the skipped bytes for `#skip` and `#skipTo` so that the caller can use them if neeed be
   * Updated `loadM8File` to use `M8FileReader` instead of a file path
   * Updated all class constructors to initialize themselves using the default values M8 would use
   * Updated `Song` to default to the appropriate `Scale` objects